refactor(core-tests) library classes renamed.
